amended interface name. extracted drink interfaces to own file
